Add Sentry config
This initialises Sentry for Rails with a few extra options.

As recommended by the Sentry docs, this configures the breadcrumbs to be
picked up from the Active Support and HTTP loggers, this enhances the
breadcrumbs we'll see in Sentry issues.

Certain URL parameters (ones containing sensitive information such as
auth tokens) are filtered out.

Some exceptions have been included, specifically ones where there has
been a client error when making the request and we wouldn't be able to
action.