Add event tracking using BigQuery
We want to be able to track the usage of the service in a way that is
privacy friendly and in-line with how we record analytics in other
services.

Using BigQuery is an established pattern across DfE so we're leaning on
the prior art to do the same here.

It is disabled in development by default and needs to be explicitly
enabled as you need to get a key from the Google Cloud console in order
to use it.

Follow the instructions in the README for more detail.