Handle missing name fields in the CSV upload
There is a scenario where someone uploads a CSV that has a blank name.

Currently, our formatting code doesn't handle this and throws an error.

I opted to add a guard clause to the formatting method to avoid the
error.

This means that the attempt to create the entry will also fail but won't
throw an error, due to the validation on the presence of the name
fields.

Potentially, this could cause confusion, as a row with a missing first
name will simply not show in the preview of the upload.