Add source_column_count validation
When creating a ChildrensBarredListEntry from a CSV row we want to
validate that the row has the correct number of columns to avoid the
data being incorrectly populated, particularly when Excel unhelpfully
removes the first column if it has no values i.e. none of the rows have
TRNs which is common.

Add a source_column_count parameter that can be passed when we want to
validate.L