Fix path helpers in support interface controller
The DSI-related before actions in this controller were erroring when
trying to access the Features section. This is because feature flags are
implemented via an engine and the engine code wasn't able to resolve the
path helpers in the main app.

Prefixing the path helpers with `main_app` addresses this.

Rather than use the shared before action in ApplicationController, we
instead:

- have the support interface controller skip `authenticate_dsi_user!`, as
  it does its own check for an internal DSI user anyway
- implement a locally defined `handle_expired_session!`

More work is needed here to refine the support user UX, but this fixes
the issue for the time being.