Fix uncheck checkbox issue
We were omitting unpermitted params because the internal checkbox was sending params as multiples.
The formbuilder library expects that a single checkbox is coded within a checkbox fieldset and that an array param is expected.
This seems like overkill when we can omit the fieldset and default to a false form param value in the controller.