Pin exact third party github action versions
Github action tags can be changed and pointed at new versions of a repo, this can lead to supply chain attacks and evidence of this occuring on other projects outside DfE exists.

Pinning to exact git hashes ensures that the version of the action cannot be changed without us explicitly changing it.